feat: scaleAndOrient 메서드 UIImage 익스텐션으로
