+++ Doesaegim/Doesaegim/Presentation/DiaryScene/FaceDetection/View/FaceDetectController.swift
@@ -177,7 +177,7 @@ extension FaceDetectController {
         pathLayer = nil
         imageView.image = nil
         
-        let correctedImage = scaleAndOrient(image: image)
+        let correctedImage = image.scaleAndOrient()
         
         imageView.image = correctedImage
         
@@ -211,86 +211,6 @@ extension FaceDetectController {
         pathLayer = drawingLayer
         viewModel?.pathLayer = pathLayer
         view.layer.addSublayer(pathLayer!)
-    }
-    
-    /// 이미지의 해상도를 재조정하고 상태에 따라 방향전환을 해주는 메서드
-    /// - Parameter image: 반영할 `UIImage`
-    /// - Returns: 해상도와 방향이 조정된 `UIImage`
-    private func scaleAndOrient(image: UIImage) -> UIImage {
-        
-        let maxResolution: CGFloat = 640
-        
-        guard let cgImage = image.cgImage else { return image }
-        
-        let width = CGFloat(cgImage.width)
-        let height = CGFloat(cgImage.height)
-        var transform = CGAffineTransform.identity
-        
-        var bounds = CGRect(x: 0, y: 0, width: width, height: height)
-        
-        let isWidthBiggerThanMaxResolution = width > maxResolution
-        let isHeightBiggerThanMaxResolution = height > maxResolution
-        if isWidthBiggerThanMaxResolution || isHeightBiggerThanMaxResolution {
-            let ratio = width / height
-            if width > height {
-                bounds.size.width = maxResolution
-                bounds.size.height = round(maxResolution / ratio)
-            } else {
-                bounds.size.width = round(maxResolution * ratio)
-                bounds.size.height = maxResolution
-            }
-        }
-        
-        let scaleRatio = bounds.size.width / width
-        let orientation = image.imageOrientation
-        switch orientation {
-        case .up:
-            transform = .identity
-        case .down:
-            transform = CGAffineTransform(translationX: width, y: height).rotated(by: .pi)
-        case .left:
-            let boundsHeight = bounds.size.height
-            bounds.size.height = bounds.size.width
-            bounds.size.width = boundsHeight
-            transform = CGAffineTransform(translationX: 0, y: width).rotated(by: 3.0 * .pi / 2.0)
-        case .right:
-            let boundsHeight = bounds.size.height
-            bounds.size.height = bounds.size.width
-            bounds.size.width = boundsHeight
-            transform = CGAffineTransform(translationX: height, y: 0).rotated(by: .pi / 2.0)
-        case .upMirrored:
-            transform = CGAffineTransform(translationX: width, y: 0).scaledBy(x: -1, y: 1)
-        case .downMirrored:
-            transform = CGAffineTransform(translationX: 0, y: height).scaledBy(x: 1, y: -1)
-        case .leftMirrored:
-            let boundsHeight = bounds.size.height
-            bounds.size.height = bounds.size.width
-            bounds.size.width = boundsHeight
-            transform = CGAffineTransform(translationX: height, y: width).scaledBy(x: -1, y: 1).rotated(by: 3.0 * .pi / 2.0)
-        case .rightMirrored:
-            let boundsHeight = bounds.size.height
-            bounds.size.height = bounds.size.width
-            bounds.size.width = boundsHeight
-            transform = CGAffineTransform(scaleX: -1, y: 1).rotated(by: .pi / 2.0)
-        default:
-            transform = .identity
-        }
-        
-        return UIGraphicsImageRenderer(size: bounds.size).image { rendererContext in
-            let context = rendererContext.cgContext
-            
-            if orientation == .right || orientation == .left {
-                context.scaleBy(x: -scaleRatio, y: scaleRatio)
-                context.translateBy(x: -height, y: 0)
-            } else {
-                context.scaleBy(x: scaleRatio, y: -scaleRatio)
-                context.translateBy(x: 0, y: -height)
-            }
-            
-            context.concatenate(transform)
-            context.draw(cgImage, in: CGRect(x: 0, y: 0, width: width, height: height))
-        }
-        
     }
     
 }
